backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Marsel <lllmarselll@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.31.1"}
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
python-dotenv = "^1.0.1"
email-validator = "^2.2.0"
mypy = "^1.11.2"
passlib = "^1.7.4"
bcrypt = "4.0.1"
types-passlib = "^1.7.7.20240819"
fastapi-pagination = "^0.12.31"
python-slugify = "^8.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
